requirements.txt
requests
psycopg[binary]   # Postgres driver
python-dotenv      # only for local testing
schedule  # for local task scheduling